B2: Move 11 gRPC main classes + services file to lucien-distributed [Luciferase-90t]
git mv (history-preserving) for all 11 classes:
- forest/ghost/grpc/ (7): GhostCommunicationManager, GhostExchangeServiceImpl,
  GhostServiceClient, MortonKeySerde, ProtobufConverters, SimpleServiceDiscovery,
  TetreeKeySerde
- balancing/grpc/ (3): BalanceCoordinatorClient, BalanceCoordinatorServer,
  GrpcBalanceExchange
- GrpcGhostChannel: forest/ghost/ -> forest/ghost/grpc/ (package relocated;
  updated package decl to forest.ghost.grpc; added imports for GhostChannel,
  GhostElement, GhostType from forest.ghost — previously implicit same-package)

META-INF/services/SpatialKeySerde: git mv lucien -> lucien-distributed (content
unchanged; serdes keep package forest.ghost.grpc). Absence from lucien prevents
ClassNotFoundException on ServiceLoader lookup.

Verification: mvn -pl lucien-distributed -am compile → 11 files, BUILD SUCCESS.
grep lucien/src/main for forest.ghost.grpc/balancing.grpc/io.grpc → empty
(only GhostExchange.java:31 javadoc {@link} — fixed in B4).
