+++ lucien-distributed/src/main/java/com/hellblazer/luciferase/lucien/forest/ghost/grpc/GhostServiceClient.java
