+++ lucien-distributed/src/main/resources/META-INF/services/com.hellblazer.luciferase.lucien.SpatialKeySerde
